requirements.txt
Python 3.8+
PyTorch
NumPy
Matplotlib